requirements.txt
langchain
google-generativeai
faiss-cpu
streamlit
requests
sentence-transformers